Refactor common functions into JobUtils trait
Since we are going to add more SparkJobs to this service, much of the
functionality can be reused from the existing implementation. We pull
out the common functions into a JobUtils trait, which can be mixed in
to any SparkJob.
